cardex2/cardex/requirements.txt
autopep8==1.5.7
dj-database-url==0.5.0
pycodestyle==2.7.0
python-decouple==3.4
Unipath==1.1
appdirs==1.4.4
asgiref==3.5.2
astroid==2.8.6
awscli==1.22.33
bcrypt==3.2.2
boto==2.49.0
boto3==1.20.33
botocore==1.23.33
certifi==2021.5.30
cffi==1.15.0
colorama==0.4.3
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
distlib==0.3.2
Django==3.2.10
django-import-export==2.8.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docutils==0.15.2
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.10
paramiko==2.11.0
Pillow==8.4.0
pipenv==2021.5.29
pipfile==0.0.2
platformdirs==2.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pylint==2.11.1
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2020.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.3.1
storage==0.0.4.3
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.0.0
tzdata==2022.1
urllib3==1.26.8
virtualenv==20.4.7
virtualenv-clone==0.5.4
wincertstore==0.2
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
django-import-export==2.8.0